feat: support plugin definitions with version number.
In order to support https:// urls for plugins we can not simply use names anymore. The https: url does not give insight on what npm package is available at the target url making caching impossible. The implementation is pretty strict, not allowing for anything but https urls and exact version definitions in order to keep the cache fast.

BREAKING CHANGE: this changes the input format definition from a set of string-separated plugins to an object structure.